Example of actually sending it chunked
But this does not fully work yet:
- on the server: 'unquoted carriage return found in data'
- on the client: 'Broken pipe'

The bad thing is that I don't know yet what's the root cause of the
issue.

Also, in the client the exceptions are wrapped, which makes quite hard
to get a useful stack trace and debug. Having exceptions and requests
library wrapped is really getting in my way.