Be resilient to attempts to cancel finished jobs
Arguably the underlying layers shall not throw exceptions for http error
status codes, as the responses can convey useful information. Talking
about pyrestcli and the BadRequestException.

Arguably we could have a hierarchy of typed exceptions for better
exception management.

Nevertheless, I just control this condition and avoid that duty to the
upper layers.